Fix Python writer string literal escaping.
This change uses repr() to output a string in which quotes and newlines are
correctly quoted. Though repr() makes Julia-compatible string literals, these
literals are also compatible with Python.

An example where this caused problems is if a file being documented contained a
function with a multi-line function as a default function argument.