Workaround parsing of constructor docstrings in Julia's AST
Julia's AST parser (as of v0.6) does not correctly parse docstrings associated
with type constructors. Unlike other functions, docstrings do not get combined
into a AST along with the function implementation and instead the string ends
up seperate as just a lone string literal.

In practice, Julia's 'help' tool will show type-level documentation when trying
to lookup a constructor. The work-around in this commit does essentially the
same thing when a constructor function is autofunction'd by substituting the
type's docstring.